Level Up screen added Player health bar added
(cherry picked from commit f7b7633d8913f75754f4b5193ade0fd1f7bb9196)

+++ Proyecto_1/Assets/Game/Scripts/EnemyStats.cs
@@ -7,7 +7,7 @@ public class EnemyStats : MonoBehaviour
     public int health = 2;
     public float speed = 3f;
     public float attackSpeed = 1f, attackDelay = 2f;
-    public int attackDamage = 1;
+    public float attackDamage = 1;
     [SerializeField] GameObject gemPrefab;
